app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
python-dotenv==1.0.1
aiosmtplib==3.0.2
httpx==0.27.2
supabase==2.10.0
python-multipart==0.0.20
